project-path-apis/api/requirements.txt
alembic
aniso8601
bcrypt
contextlib2
coverage
cryptography
Flask
Flask-API
Flask-Bcrypt
Flask-Cors
Flask-JWT-Extended
flask-marshmallow
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
gevent
greenlet
gunicorn
marshmallow
marshmallow-enum
marshmallow-sqlalchemy
psycopg2-binary
PyJWT
python-dateutil
python-dotenv
SQLAlchemy
SQLAlchemy-Utils
google-auth
google-auth-httplib2
google-api-python-client
sendgrid
celery
redis
PyPDF2
boto3
pgvector
transformers==4.36.0
openai
sentence-transformers
